Code review: Fix compile warnings in clsFSSrcSession
git-svn-id: http://192.168.20.11/sequoiadb/trunk@32156 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsFSSrcSession.cpp
@@ -517,7 +517,7 @@ namespace engine
             rc = SDB_INVALIDARG ;
             goto error ;
          }
-         if ( len < ele.valuestrsize() )
+         if ( len < (UINT32)ele.valuestrsize() )
          {
             PD_LOG( PDWARNING, "Session[%s]: CS Name[%s] is too long",
                     sessionName(), ele.valuestr() ) ;
@@ -558,7 +558,7 @@ namespace engine
             rc = SDB_INVALIDARG ;
             goto error ;
          }
-         if ( len < ele.valuestrsize() )
+         if ( len < (UINT32)ele.valuestrsize() )
          {
             PD_LOG( PDWARNING, "Session[%s]: CL Name[%s] is too long",
                     sessionName(), ele.valuestr() ) ;
